+++ firmware/main/uni_hid_device.h
@@ -19,6 +19,7 @@ limitations under the License.
 #ifndef UNI_HID_DEVICE_H
 #define UNI_HID_DEVICE_H
 
+#include <stdbool.h>
 #include <stdint.h>
 
 #include "btstack.h"
@@ -92,21 +93,21 @@ void uni_hid_device_remove_entry_with_channel(uint16_t channel);
 
 void uni_hid_device_request_inquire(void);
 
-void uni_hid_device_set_disconnected(uni_hid_device_t* d);
+void uni_hid_device_set_connected(uni_hid_device_t* d, bool connected);
 
 void uni_hid_device_set_cod(uni_hid_device_t* d, uint32_t cod);
-uint8_t uni_hid_device_is_cod_supported(uint32_t cod);
+bool uni_hid_device_is_cod_supported(uint32_t cod);
 
 void uni_hid_device_set_hid_descriptor(uni_hid_device_t* d,
                                        const uint8_t* descriptor, int len);
-uint8_t uni_hid_device_has_hid_descriptor(uni_hid_device_t* d);
+bool uni_hid_device_has_hid_descriptor(uni_hid_device_t* d);
 
-void uni_hid_device_set_incoming(uni_hid_device_t* d, uint8_t incoming);
-uint8_t uni_hid_device_is_incoming(uni_hid_device_t* d);
+void uni_hid_device_set_incoming(uni_hid_device_t* d, bool incoming);
+bool uni_hid_device_is_incoming(uni_hid_device_t* d);
 
 void uni_hid_device_set_name(uni_hid_device_t* d, const uint8_t* name,
                              int name_len);
-uint8_t uni_hid_device_has_name(uni_hid_device_t* d);
+bool uni_hid_device_has_name(uni_hid_device_t* d);
 
 void uni_hid_device_set_product_id(uni_hid_device_t* d, uint16_t product_id);
 uint16_t uni_hid_device_get_product_id(uni_hid_device_t* d);
@@ -117,15 +118,17 @@ uint16_t uni_hid_device_get_vendor_id(uni_hid_device_t* d);
 void uni_hid_device_dump_device(uni_hid_device_t* d);
 void uni_hid_device_dump_all(void);
 
-uint8_t uni_hid_device_is_orphan(uni_hid_device_t* d);
+bool uni_hid_device_is_orphan(uni_hid_device_t* d);
 
 void uni_hid_device_guess_controller_type(uni_hid_device_t* d);
-uint8_t uni_hid_device_has_controller_type(uni_hid_device_t* d);
+bool uni_hid_device_has_controller_type(uni_hid_device_t* d);
 
 void uni_hid_device_process_gamepad(uni_hid_device_t* d);
 
 void uni_hid_device_set_connection_handle(uni_hid_device_t* d,
                                           hci_con_handle_t handle);
+
+void uni_hid_device_set_state(uni_hid_device_t* d, enum DEVICE_STATE s);
 
 // events